requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cx-Freeze==6.15.12
cx-Logging==3.1.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
lief==0.13.2
numpy==1.26.2
pandas==2.1.4
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
